Solution for Part 2 of 8th of december

+++ 8December/solution.h
@@ -1,5 +1,6 @@
 #ifndef SOLUTION8_H
 #define SOLUTION8_H
+#include <bits/stdc++.h>
 size_t solution8();
-int second_solution8();
+size_t second_solution8();
 #endif
